AP CS medium problems
These problems are based on free response questions from old AP CS-A exams, 2004 and later. Some are from the exams and some were just created to have a similar difficulty. These problems are good practice for the logic/loop/array aspects of the exam. See also the Basic Recursion section and the official college board CS-A page.